Table from mui

+++ src/components/User/UserInfo.tsx
@@ -1,8 +1,49 @@
+import { TableBody, TableCell, TableHead, TableRow } from "@mui/material";
 import { tableHeader } from "../../models/constants";
 import { FormValueType, TableHeaderType } from "../../models/types";
 import { Card } from "../UI/Card/Card";
+import Table from "@mui/material/Table";
+
+// const MyTable = () => {
+//   const columns = [{ id: "name", label: "Name" }];
+//   const rows = [{ name: "haha" }];
+//   return (
+//     <Table stickyHeader aria-label="sticky table">
+//       <TableHead>
+//         <TableRow>
+//           {columns.map((column) => (
+//             <TableCell
+//               key={column.id}
+//               // align={column.align}
+//               // style={{ minWidth: column.minWidth }}
+//             >
+//               {column.label}
+//             </TableCell>
+//           ))}
+//         </TableRow>
+//       </TableHead>
+//       <TableBody>
+//         {rows
+//           // .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
+//           .map((row: any) => {
+//             return (
+//               <TableRow hover role="checkbox" tabIndex={-1}>
+//                 {columns.map((column: any) => {
+//                   const value = row[column.id];
+//                   return <TableCell key={column.id}>{value}</TableCell>;
+//                 })}
+//               </TableRow>
+//             );
+//           })}
+//       </TableBody>
+//     </Table>
+//   );
+// };
 
 export const UserInfo = (props: any) => {
+  // const columns = [{ id: "name", label: "Name" }];
+  // const rows = [{ name: "haha" }];
+
   const deleteHandler = (item: FormValueType) => {
     props.onDelete(item);
   };
@@ -12,40 +53,57 @@ export const UserInfo = (props: any) => {
   };
 
   const cellDataJSX = (colHeader: TableHeaderType, userInfo: FormValueType) => {
-    if(colHeader.id === 'profilePic') {
-      return <img className="w-20 h-20 rounded-xl mx-auto object-contain" src={userInfo[colHeader.id].split("#")[1]} alt="userImg"></img>
+    if (colHeader.id === "profilePic") {
+      return (
+        <img
+          className="w-20 h-20 rounded-xl mx-auto object-contain"
+          src={userInfo[colHeader.id].split("#")[1]}
+          alt="userImg"
+        ></img>
+      );
+    } else if (colHeader.id === "delete-action") {
+      return <button onClick={() => deleteHandler(userInfo)}>delete</button>;
     } else {
-      return <span className="truncate">{userInfo[colHeader.id]}</span>
+      return <span className="truncate">{userInfo[colHeader.id]}</span>;
     }
-  } 
+  };
 
   return (
-    <Card>
-      <div className="flex flex-row">
-        {tableHeader.map((colHeader: TableHeaderType) => (
-          <div key={`header-${colHeader.id}`} className="w-40 truncate">
-            {colHeader.colName}
-          </div>
-        ))}
-      </div>
-      <div className="flex flex-col">
-        {props.userDetails.map((userInfo: FormValueType) => (
-          <div
-            key={`row-${userInfo.id}`}
-            className="flex flex-row cursor-pointer"
-            onDoubleClick={() => editHandler(userInfo)}
-          >
-            {tableHeader.map((colHeader: TableHeaderType) => (
-              <div className="flex w-40 items-center" key={`cell-${colHeader.id}`}>
-                {colHeader.id === "delete-action" && (
-                  <button onClick={() => deleteHandler(userInfo)}>delete</button>
-                )}
-                {colHeader.id !== "delete-action" && cellDataJSX(colHeader, userInfo)}
-              </div>
+    <>
+      <Card>
+      <Table stickyHeader aria-label="sticky table">
+        <TableHead>
+          <TableRow>
+            {tableHeader.map((column: TableHeaderType) => (
+              <TableCell
+                key={column.id}
+                // align={column.align}
+                // style={{ minWidth: column.minWidth }}
+              >
+                {column.colName}
+              </TableCell>
             ))}
-          </div>
-        ))}
-      </div>
-    </Card>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          {props.userDetails
+            .map((row: FormValueType) => {
+              return (
+                <TableRow hover onDoubleClick={() => editHandler(row)} key={`row-${row.id}`}>
+                  {tableHeader.map((column: TableHeaderType) => {
+                    return (
+                      <TableCell key={column.id} align="center">
+                        {cellDataJSX(column, row)}
+                      </TableCell>
+                    );
+                  })}
+                </TableRow>
+              );
+            })}
+        </TableBody>
+      </Table>
+      </Card>
+
+    </>
   );
 };
